refractor(post): 서비스 분리
